Move punnett_2x2 to punnett_generator.py
